+++ package.json
@@ -19,11 +19,6 @@
     "vega-lite": "^5.2.0"
   },
   "devDependencies": {
-    "@types/jest": "^27.5.1",
-    "@types/jsdom": "^16.2.14",
-    "jest": "^28.1.0",
-    "jest-environment-jsdom": "^28.1.0",
-    "ts-jest": "^28.0.2",
     "typescript": "^4.6.4",
     "webpack": "^5.72.1",
     "webpack-cli": "^4.9.2"
