+++ packages/adapters/src/README.md
