+++ packages/adapters/src/VegaLiteAdapter.ts
