+++ packages/core/src/Structure/README.md
