+++ packages/core/webpack.config.js
